+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `HorizontalProductSummary` prop to the `autocomplete-result-list.v2`.
+
 ## [2.4.0] - 2021-02-01
 
 ### Added